LatinIME: Update all dicts for ICS
In ICS the binary dictionary format for dictionaries
has changed. I decompiled them all to xml and then
recompiled with ICS makedict. They will still not be
considered installed however. I have confirmed the
files work by messing with LatinImeDictionaryPack.apk
though, a part of the gapps.

I'm just submitting this in the hope that someone
else will also try to figure it out.

It also decreases the size of LatinIME.apk slightly
as the new format results in greater compression.

Change-Id: Ia4d1305bf64e6529cd90b44aa9af39dcefae1282